requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
astroid==2.12.8; python_full_version >= '3.7.2'
autopep8==1.7.0
backports.zoneinfo==0.2.1; python_version < '3.9'
dill==0.3.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
dj-database-url==1.0.0
django==4.0.5
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
isort==5.10.1; python_version < '4.0' and python_full_version >= '3.6.1'
lazy-object-proxy==1.7.1; python_version >= '3.6'
mccabe==0.7.0; python_version >= '3.6'
pillow==9.2.0
platformdirs==2.5.2; python_version >= '3.7'
psycopg2==2.9.3; python_version >= '3.6'
pycodestyle==2.9.1; python_version >= '3.6'
pyjwt==2.4.0; python_version >= '3.6'
pylint==2.15.0
pytz==2022.2.1
setuptools==65.3.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version < '3.11'
tomlkit==0.11.4; python_version >= '3.6' and python_version < '4.0'
typing-extensions==4.3.0; python_version < '3.10'
whitenoise==6.2.0; python_version >= '3.7'
wrapt==1.14.1; python_version < '3.11'
